requirements.txt
numpy==1.26.4
scikit_learn==1.4.1.post1
pandas==2.2.1
joblib==1.3.2
flake8==7.0.0
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
requests==2.31.0
psycopg2-binary==2.9.9
streamlit==1.32.2
asyncpg==0.29.0
python-dotenv==1.0.1
jupyter==1.0.0
<<<<<<< HEAD
apache-airflow==2.7.2
=======
>>>>>>> 918a146816d2cb94bff7d8b184e2bc15677387e0
